Refactor basic routines

+++ src/hal_hisi.h
@@ -1,6 +1,7 @@
 #ifndef HAL_HISI_H
 #define HAL_HISI_H
 
+#include <math.h>
 #include <stdbool.h>
 #include <sys/types.h>
 
@@ -18,7 +19,7 @@
 int hisi_SYS_DRV_GetChipId();
 const char *hisi_cv100_get_mii_mux();
 bool hisi_ev300_get_die_id(char *buf, ssize_t len);
-int hisi_get_temp();
+float hisi_get_temp();
 
 int sony_ssp_read_register(int fd, unsigned char i2c_addr,
                            unsigned int reg_addr, unsigned int reg_width,
